+++ packages/miniapp-compile-config/src/plugins/AutoInstallNpm.js
@@ -1,6 +1,37 @@
 const { autoInstallNpm } = require('miniapp-builder-shared');
-const { writeJSONSync } = require('fs-extra');
+const { writeJSONSync, existsSync, readJSONSync } = require('fs-extra');
 const { join, dirname } = require('path');
+
+function isObjectValueEqual(left, right) {
+  var leftProps = Object.getOwnPropertyNames(left);
+  var rightProps = Object.getOwnPropertyNames(right);
+
+  if (leftProps.length !== rightProps.length) {
+    return false;
+  }
+
+  for (var i = 0; i < leftProps.length; i++) {
+    var propName = leftProps[i];
+    var propLeft = left[propName];
+    var propRight = right[propName];
+    if (propLeft !== propRight) {
+      return false;
+    }
+  }
+  return true;
+}
+
+function shouldWritePackageJson(packageJsonPath, currentDependencies) {
+  if (!currentDependencies) {
+    return false;
+  }
+  if (!existsSync(packageJsonPath)) {
+    return true;
+  }
+  const oldDependencies = readJSONSync(packageJsonPath).dependencies;
+  return !isObjectValueEqual(currentDependencies, oldDependencies);
+}
+
 /**
  * Auto install npm
  */
@@ -14,19 +45,24 @@ module.exports = class AutoInstallNpmPlugin {
     compiler.hooks.done.tapAsync('AutoInstallNpmPlugin', async(stats, callback) => {
       const packageJsonFilePath = [];
       const distDir = stats.compilation.outputOptions.path;
+
       // Generate package.json
-      if (this.dependencies) {
-        writeJSONSync(join(distDir, 'package.json'), { dependencies: this.dependencies });
+      const writePath = join(distDir, 'package.json');
+      if (shouldWritePackageJson(writePath, this.dependencies)) {
+        writeJSONSync(writePath, { dependencies: this.dependencies });
         packageJsonFilePath.push('');
       }
       if (this.subPackages) {
         this.subPackages.forEach(({ dependencies = {}, source = '' }) => {
-          writeJSONSync(join(distDir, dirname(source), 'package.json'), { dependencies });
-          packageJsonFilePath.push(dirname(source));
+          const writePath = join(distDir, dirname(source), 'package.json');
+          if (shouldWritePackageJson(writePath, dependencies)) {
+            writeJSONSync(writePath, { dependencies });
+            packageJsonFilePath.push(dirname(source));
+          }
         });
       }
 
-      if (!this.autoInstall) {
+      if (!this.autoInstall || packageJsonFilePath.length === 0) {
         return callback();
       }
       await autoInstallNpm(callback, { distDir, packageJsonFilePath });