+++ packages/miniapp-webview-config/src/plugin/platforms/ali.ts
@@ -1,0 +1,3 @@
+export const templateXml = `
+<web-view src="{{url}}{{params}}" onMessage="{{onMessage}}"  onLoad="{{onWebviewLoad}}" onError="{{onWebviewError}}"></web-view>
+`;