+++ packages/jsx2mp-loader/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.4.34] - 2022-01-06
+
+### Fixed
+
+- update the logic of replacing rem to rpx
+
+## [0.4.33] - 2021-12-14
+
+### Changed
+
+- update miniapp-builder-shared to v3.0.0
+
 ## [0.4.32] - 2021-11-03
 
 ### Fixed