+++ packages/miniapp-history/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.1.7] - 2022-01-12
+
+### Changed
+
+- output es5 code in `lib` dir
+
+## [0.1.6] - 2021-12-14
+
+### Changed
+
+- `lib` dir is `esm` result
+
 ## [0.1.5] - 2021-07-20
 
 ### Added